pyproject.toml
[project]
name = "kis-trading-bot"
version = "0.1.0"
description = "KIS 3x ETF bot with StochRSI K/D slices"
authors = [{name="Mobum Shin"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2",
  "pyyaml>=6",
  "typer>=0.12",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "httpx>=0.27",
  "websockets>=12",
  "pandas>=2.2",
]

[project.scripts]
kisbot = "kisbot.main:app"

[tool.alembic]
script_location = "alembic"
